+++ ochk/utils.go
@@ -6,9 +6,9 @@ import (
 )
 
 func generateRandName() string {
-	return fmt.Sprintf("tf-acc-test-%s", acctest.RandStringFromCharSet(8, acctest.CharSetAlphaNum))
+	return generateShortRandName()
 }
 
 func generateShortRandName() string {
-	return fmt.Sprintf("tf-%s", acctest.RandStringFromCharSet(8, acctest.CharSetAlphaNum))
+	return fmt.Sprintf("tf-%s", acctest.RandStringFromCharSet(4, acctest.CharSetAlphaNum))
 }